+++ python/GafferTest/TestCase.py
@@ -348,7 +348,7 @@ class TestCase( unittest.TestCase ) :
 
 		subprocess.check_call( [ "gaffer", "python", script ] )
 
-	def assertEqualToFloatPrecision( self, value0, value1 ) :
+	def assertFloat32Equal( self, value0, value1 ) :
 
-		from GafferTest import equalToFloatPrecision
-		return equalToFloatPrecision( value0, value1 )
+		from GafferTest import float32Equal
+		return float32Equal( value0, value1 )